pkg/pgutil: add docs and remove `Item.Scan()` callback
* pkg/pgutil: add docs to public types, functions, fields, etc
* pkg/pgutil: remove `Item.Scan()` callback (redundant via `sql.Scanner`)
* pkg/pgutil: move `Result` type and method defs to result.go
* pkg/pguserstore: make `userEntry.Scan()` match new `Item.Scan()` signature
* pkg/pgtokenstore: make `userEntry.Scan()` match new `Item.Scan()` signature
* pkg/pgtokenstore: change `expires` column type `TIMESTAMP`->`TIMESTAMPTZ`
* pkg/types: `Token.Compare()`: `Expires` field uses `==` -> `Time.Equal()`
